python module: optimize linkage test to avoid distutils
While this introspection script still does use a lot of distutils in
order to detect Debian, we can get a little closer to removing our
reliance on it by short-circuiting the check to see if we need
-lpython3.X -- current versions of python mandate that extensions never
link to libpython.

FIXME: totally broken because this is only mandated on Unix.

... and exempted on Android. And Cygwin. And maybe MSYS2?

Okay, what the heck is the upstream solution here?

Also, hmm... isn't pkg-config already supposed to list this info? Or
python-config. But, our system dependency (probes sysconfig) does not
know that information, so how do we find it???