switch an mlog.warning of ambiguous purpose to FeatureDeprecated
It probably makes sense to be a deprecation rather than just a warning,
because we don't actually expect this to be there for any reason other
than an attempt at `copy: true`.

Using the new support for location nodes in Feature checks, we can
continue to print the location of this configured file. We do drop
mention of the input filename, but the line number should work just fine
for locating the cause.