first pass at migrating to dataclasses
To start with, all init function arguments are converted to dataclass
decorated attributes without further changes.

In some cases, this requires renaming named init variables to class
instance variables when they have different names. In other cases it
requires setting the default to [] if defaulting to none... or if None
was manually passed in.

TODO: the former case can be fixed by renaming all callers too (cf. at a
minimum unstable_external_project) and verifying that None cannot ever
actually be passed, but is only being used as a sentinel for
"unspecified".