python module: when falling back to unversioned name, check version
find_installation('python2') and find_installation('python3') can fall
back to detecting 'python' and running with that. This is supposed to
check the version, but that was broken because the sanity check only
sees 'python' and cannot tell what version to try to enforce.

Pass the real name in this fallback mode, and use it in the version
check as the preferred identity.