mtest: fix rebuilding all before running tests
Regression in the original implementation of commit
79e2c52a15e896e46ff3cfa3ec16fbf3f132ee01.

If an explicit list of targets is passed on the CLI, then that is passed
to rebuild_deps. If not, we pass every loaded test to rebuild_deps
instead. This means we cannot distinguish between "trying to run all
tests" and "trying to run specific tests". We then load all the deps for
all tests, and try to build them all as explicit arguments to the
underlying ninja.

This is usually papered over by people running `ninja test`, which
depends on "all" anyway as a prerequisite for invoking the underlying
`meson test --no-rebuild --print-errorlogs`, however, instead running
`meson setup builddir && meson test -C builddir` with under-specified
test deps would actually build a handful of deps and not anything else,
then fail.

Instead, pass no tests to rebuild_deps, so that we iterate over nothing
and add no explicit targets. Hence, meson rebuilds the default "all"
rule.