coverage: cheat the gcovr heuristics to prevent wasted guesswork
What and how gcovr works is best described as... confusing. But it turns
out that it tries to repeatedly run gcov until it figures out which
workdir the compiler was originally run in, and it does this by walking
the build tree, trying the source tree, etc.

--object-directory overrides that, although strictly speaking gcov/r only
cares about *.gc{da,no} files and not objects anyway; also, it's not
related to the --object-directory option of gcov itself.

Basically, don't think about it, just accept that this option somehow
makes things faster. Sigh.