python module: stop using distutils "embed" probe on sufficiently new python
On python >=3.8, this information is... ***mostly*** encoded in the
sysconfig vars.

In distutils, it is always necessary to link to libpython on Windows;
for posix platforms, it depends on the value of LIBPYTHON (which is the
library to link to, possibly the empty string) as generated by
configure.ac and embedded into python.pc and python-config.sh, and then
coded a second time in the distutils python sources.

There is a caveat -- python.pc and python-config.sh disagree with
distutils when python is not built shared. In that case, distutils
*never* links to libpython, while the former act the same as a shared
build.

FIXME: we need to resolve that before this commit makes sense to
include. I will open a CPython ticket to discuss the inconsistency.

This is the other half of the fix for ticket 7702