consolidate knowledge of build_target() implementations
Dispatching to the correct function is a giant list of cases.  But the
function table already has this information mapped based on the
target_type.

Handling typed arguments becomes a bit easier, since we can just
dispatch to the "real" function's argument checking. This improves the
error message when passing structured sources to jar, which is forbidden
by `jar()` but permitted by `build_target(..., target_type: 'jar')`
until it hits build.py, at which point the Jar constructor specifically
checks for it and errors out.

For kwargs, we can filter out the arguments that are valid for
build_target, but not for the real function.