do not treat empty directories as initialized subprojects
When meson sees a subproject's name exists on disk, it avoids doing
further work, either respecting what is there or erroring out if it's a
file which we cannot (should not) overwrite. This is generally correct,
but there's an edge case where a subproject is both a `*.wrap` and a git
submodule. It is then possible to end up with a project tarball by way
of git-archive (helpfully broken on users' behalf by software forges
everywhere, that don't provide a way to disable this for repositories
where it's known to not work). This tarball has a placeholder directory,
an *empty* directory, which git-archive produces for unclear reasons
that don't produce useful behavior. Meson then chokes on it, and says:

ERROR: Subproject exists but has no meson.build file

Work around this case by treating a totally empty directory as
tantamount to nonexistent. There's nothing to overwrite, and it's
probably this git edge case.