clang: do not force color to "always" when auto is set
GCC accepts an argument -fdiagnostics-color=WHEN, and
-fdiagnostics-color without WHEN is interpreted as "always". The default
defined when gcc itself is compiled, for when no flag is specified, is
probably auto. We provide all three options.

But clang doesn't provide this granularity at all. It only has an always
flag and a never flag, and when specifying no flag it acts like auto.
This was probably assumed to not matter, since people can just not
specify it, but the GCC behavior is useful when overriding flags that
came from elsewhere.

For some reason, the initial implementation of -Db_colorout ignored this
logic entirely, and insisted on passing a clang option no matter what --
specifically, it passes the "always" option when "auto" is requested.
Simply delete this and pass [] instead, which makes clang do exactly
what the user asked for.