i18n module: be broadly supportive of portable gettext tools
There are a number of implementations for the gettext utility `msgfmt`,
supporting various options. The simplest, and most common, use case is
to compile .po files into .mo files, and this should be able to work on
gettext implementations other than the GNU one.

In particular, Solaris has an implementation of the .mo workflow, and
the next version of POSIX standardizes this common subset of the
`msgfmt` utility. See https://www.austingroupbugs.net/view.php?id=1122

The problem is that we were passing some pretty portable arguments in an
unportable manner. The `-o` option-argument and its associated argument
came after the input file operand, which violates the POSIX Utility
Syntax Guidelines, and happens to not be supported by Solaris gettext.
The GNU gettext doesn't care; GNU invented GNU argument permutation.

Switch the order around so that our use is POSIX-conformant and thus
portable.