CI: fix macOS disaster with Github Actions-based python yet again
We need this in the unittests job as well, which manifested the same
issue as the project tests. Something in homebrew depends on python,
which yanks that in, which means we have the same issue we fixed in
commit 0e5d632a62ef79b29df98a437c02d02419eaf8e4.

I have not seen the Qt4macos test fail in the same way, and it only uses
pkgconfig, ninja, and gcc. Hopefully this usage of homebrew is light
enough we won't have an issue. This may be foolishly optimistic of me,
but time will tell. I'd rather not spend the extra CI time if it's not
actually required.