micro-optimize iteration of evaluated set
It is generally accepted practice to convert dict.keys() to a list
before iterating over it and e.g. deleting values, as keys returns a
live-action view. In this case, we use the difference of *two* dict
keys, which returns a regular non-view set and doesn't need protecting.

Iteration order doesn't matter (the set already randomizes it anyway).
Avoid the cost of converting to a list.