tests: correctly handle the skippability of wxwidgets framework
Back when we originally added this "if found, do fundamental logic" code
in commit 24aa7f300a116c955b2417339ff31bee024bc55d, conditionally
performing the test logic was the state of the art. Later that year, we
implemented the MESON_SKIP_TEST protocol for actually marking skipped
tests as skipped, rather than silently doing nothing. We should use this
so that people know when the test was skipped. :)