fix broken implementation of PCH dependencies
Regression in the initial implementation of 4860. In commit
2044bc53b3534e0e91d75b5435792e4b4d342847 we fixed PCH to depend on
generated headers at all. But we did this by making them regular rebuild
dependencies, even though it should have been done via order deps, the
same way regular source headers are. If the header actually changes, and
it is *used* by the PCH file, compiler depfile support causes a rebuild.

We don't want to rebuild every time any header changes, because not all
headers are actually included in the PCH file. We just need to make sure
they exist at all in a clean build as we don't know which files are true
dependencies.