ci: upgrade github actions runner to use ubuntu-latest all the time
It was pinning an old ubuntu version for no reason I can see, and that
meant that github occasionally complains we are using something that is
deprecated and will disappear soon.

It does require a bit of mild fixup. Github slimmed down their image a
bit, so glib2 devel components aren't installed by default.