mdist: fix failure to create tar files the user asked to create
Python added a "feature" to assume tarfile extraction is meant solely
for "data", i.e. it ignores many useful features of tar such as
symlinks, ownership, or permission modes that are uncommon on Windows.
Revert this entirely, as Meson is a "fully trusted" application. It can
already execute arbitrary programs, tar files are not vulnerabilities.

In theory "tar" mode exists and is not "data", but we are fully trusted
so why split hairs?